+++ registro.h
@@ -12,7 +12,7 @@
     REGISTRO* criarRegistro(char removido, int tamanhoRegistro, long prox, int id, int idade, int tamNomeJogador, char *nomeJogador, int tamNacionalidade, char *nacionalidade, int tamNomeClube, char *nomeClube);
     char get_removido(REGISTRO *registro);
     int get_tamanhoRegistro(REGISTRO *registro);
-    long get_prox(REGISTRO *registro);
+    long long int get_prox(REGISTRO *registro);
     int get_id(REGISTRO *registro);
     int get_idade(REGISTRO *registro);
     int get_tamNomeJogador(REGISTRO *registro);
@@ -21,7 +21,7 @@
     char* get_nacionalidade(REGISTRO *registro);
     int get_tamNomeClube(REGISTRO *registro);
     char* get_nomeClube(REGISTRO *registro);
-    void set_prox(REGISTRO *registro, long prox);
+    void set_prox(REGISTRO *registro, long long int prox);
     void set_removido(REGISTRO *registro, int removido);
     void set_tamanhoRegistro(REGISTRO *registro, int tamanhoRegistro);
     void set_id(REGISTRO *registro, int id);
